requirements.txt
Django~=2.1
django-taggit
django-crispy-forms